Start work in adding a PLY-based IDL parser.
Instead of using a manual syntax parser, this adds a PLY-based one.
PLY is a Python implementation of YACC, a general syntax parser.  This
allows us to avoid a lot of complicated manual parsing and will reduce
the amount of code as we add more features to it.

This adds the new parser beside the old one since this doesn't
implement everything yet.  This only supports parsing empty
dictionaries.

Change-Id: Ic328e805ed9bf002f1c7b0386a251613b4fb9641