Fix seek slider's handling of large start times.
Previously, we were storing the start and end times of the presentation
inside a UISlider. Apparently, UISliders don't handle very large values
well, because this lead to the minimum and maximum values being
incorrect for livestreams that had a very large start time.
This changes the seek bar to store the presentation start as a double,
and simply translate the time values as they enter and leave.

Fixes: 119566312
Change-Id: I54621e6b547e643a779cfdadb99860ce1f64df16