Allow using dynamic libraries in EME plugins.
This now will copy the dynamic library or framework into the iOS
bundle so the app doesn't need to worry about loading it themselves.
This also copies the dynamic library to the output directory for
desktop to allow for easier testing, but the app is still responsible
for copying it to their install location.

Change-Id: Id7f6379ecf1fd3aeb0f99dc510c126b4b8b962a7