Fix optional/variant constructors.
This adds a filter to the general U&& constructor to ensure it doesn't
catch optional<T> or variant<...> types.  This avoids compiler errors
when using these types since it  should use the other constructors to
get the inner values.

This also adds a get move constructor to allow getting values from
move-only types.

Change-Id: I6e35721d13d3f375ea3f835fb8bf1095dbc1ca75