Move JavaScript object interactions to helper.
This creates a new helper to interact with a JavaScript-defined object.
This allows us to share the bulk of the code between the Player and
Storage objects.

This also changes the Player methods so they can be called from the
main thread, allowing them to be called from the event listeners.

Change-Id: I95f80c28e464478db1370bcf8e6778f9d53db97a