Refactor error handling.
First, this adds an Objective-C version of the Error type to hold error
info.  This also changes the callbacks to use this instead of a string.

Second, this adds a client interface in Objective-C for errors and
buffering callbacks.  This changes the load/unload methods to give
the error in the block instead of calling the callback to ensure
the block knows if an error happened.

Lastly, this removes the "type" from the C++ error.  This is basically
just two different values.  The third value will rarely be seen and
will likely never be handled specially by an app.  So it is simpler to
just report Shaka errors with the severity/category/code fields, and
native errors with those set to 0.

Closes #7
Closes #10
Closes #53

Abi-Breaking: Changing Error interface.
Change-Id: I7b2d344f1322a8b6ecb0ce8ef276fb1842870010