Allow move-only types in IDL types.
Now the Struct type that is generated can contain move-only fields so
long as they are converted to copyable types for the public interface.
This will be used by the ByteBuffer type, which is move-only.

This also makes the public methods return a const reference to the
values to avoid copying.

Abi-Breaking: Changing generated IDL type return values.
Change-Id: I2cb18a927a4d37d16859be9987438c84c9f3c309