Reduce size of variant<> type.
First, the union should use the smallest type instead of a pointer to
avoid being larger than the type it is stored with.  Second, we don't
need a size_t to store the index since it will only be as large as the
number of types.

This reduces the size of a "variant<char, char>" from 16 bytes to 2
(on a 64-bit build).  Also, by avoiding pointers and large integers,
this should reduce the alignment requirement, allowing them to be
packed more.

Abi-Breaking: Changing layout of variant type.
Change-Id: I94fe63bc05c381bfec390cb76baa8d997f5ee439