Fix iOS bundle versions.
This adds a CFBundleShortVersionString key, which is required by the
App Store.  This also auto-generates the version strings from the
current tag to match the version in the library.

This uses the same version number for FFmpeg since we need a custom
version with some patches so different versions of FFmpeg may not be
able to be used by us.

Issue #63

Change-Id: Ic808de8738d25060c8705dca90a28dbe973d3822