Have FFmpeg as a sibling framework.
The App Store doesn't allow umbrella frameworks on iOS, so we cannot
embed FFmpeg in our Framework like we wanted.  So now FFmpeg will be
a sibling framework and the app will need to embed that also.

This also updates the README and tutorial to indicate the FFmpeg
bundle is needed in the app.

Fixes #63

Change-Id: I9d572299a4a8caf6b4fe678f5828100c862f00f3