Avoid deadlock in XMLHttpRequest.
When we raise progress events, it is on the JS main thread and only
accesses JS properties, so we don't need to acquire the lock.

We synchronously fire JS events, which may call methods on this type
or may call Abort, which will call into the NetworkThread and can cause
a deadlock with that.

Issue #65

Change-Id: I1135d049eea5623bef4bcca503907198377273e9