Move FrameBuffer code to StreamBase.
Now the public type is implemented based on the code from FrameBuffer.
This replaces all the usages of the old FrameBuffer to use the new API.
Unfortunately since too much was changed and there was a file with the
same name already, git doesn't detect the file as moved.

Issue #60

Change-Id: I9b95712bf2aaa934e00e8b07008681376832009d