Add tests for the Demuxer.
This stores info about the expected frames next to each file, and uses
that to determine which frames are expected.  This allows us to test
multiple appends without duplicating frame hashes or requiring a
specific order to add to.

This uses ffprobe to determine the hashes and other info.  There is a
script to generate this data, but the data is checked into the repo to
ensure we use golden data and avoid a bug in the script causing the
tests to not run.  This allows partial manual review of the extra data.

Issue #60

Change-Id: Ic90cbcf3cdf186e479b53053acb2ded9f9ec0850