Allow removing media components.
Now most of the media pipeline can be removed from the builds.  This
allows for smaller builds when apps will always be providing another
implementation.

Now FFmpeg can be removed from the build by passing --no-demuxer,
--no-decoder, and --no-sdl-audio (and providing your own demuxer,
decoder, and audio renderer). FFmpeg is still needed for the audio
renderer to resample and handle A/V sync.

Issue #60

Change-Id: I62b05d8412d2f5dae46b3544e9397038bc10a8cf