Reduce size and alignment of optional<T>
Instead of using a pointer as the dummy value, use a "char" so the type
is smaller.  If T is a bool, the pointer would be much larger than the
type we are storing; by using a char, we don't need to store the extra
data.  This also reduces the alignment requirements when storing smaller
types.

Change-Id: I1b539cd3071abece306d79f910a8b787e0e47ee1