Make V8HeapTracer subclass HeapTracer.
This is more clear since there will be one HeapTracer object.  This
also allows accessing the V8HeapTracer while tracing, as will be
required in a follow-up.  V8HeapTracer will also store the trace list
in the prologue so we can detect when we are done in a follow-up.

This also renames TraceCommon to TraceAll to make it a bit more clear.

Change-Id: If5acc2969cc6790bbe0e90eca822a85fe9c7a726