Evict old decoded frames.
Now the default MediaPlayer will evict old DecodedFrame objects as part
of decoding.  This ensures we don't buffer them forever and makes the
job easier for the renderers.

This also moves the check for if we've buffered enough higher to avoid
extra work if we are going to sleep.

Issue #60

Change-Id: I16677277fd20b0674bf5d49926cb31ad2ab78306