Move some frame properties to StreamInfo.
Properties like width/height/etc are constant in a stream and may be
needed to initialize a decoder.  So this moves them to the StreamInfo
object.  This also adds the StreamInfo instance to the DecodedFrame so
full-frames can know what streams they came from.

Issue #60

Abi-Breaking: Changes properties of Frame types.
Change-Id: I17468656694359913c9f5eb2e042bd7cd8565c13