Add TextTrackList and HTMLMediaElement.
This adds a separate TextTrackList type since src= requires that list
to be an event listener.  This also moves the bulk of the video code
to HTMLMediaElement, which is required for src=.

This renames the file to video_element_2 so we maintain the git
history and the file is detected as moved.

Issue #60
Issue #89

Change-Id: Ic1c59208183e9dd501494aad0a49d4b263ca3b4a