Generalize TextTrackList and use same instance.
Now the TextTrackList will get updates from the MediaPlayer so the same
instance will see when new tracks are added.  It also fires the JS
events when tracks are added/removed.

This also generalizes the list so we can store other track types.  This
will make it easier to add audio/video tracks in a follow-up.

Issue #89

Abi-Breaking: Adds events to MediaPlayer::Client.
Change-Id: I41255ee80d8dd30bca9610a2f5a188434af07593