Add to_string for media enums and debug printing.
This adds a DebugPrint method to print the contents of a stream.  This
will help debugging buffers and can be called while debugging (e.g. with
gdb).  This also adds disabled code to print out the current media
state in the MSE player.  This allows us to easily enable the code to
log the current playback information.

Change-Id: I4062df1827b942dd502e230a59eb04650a7af835