Fix finding frames with KeyFrameBefore.
This wouldn't return the correct frame since it would always detect it
as in the future and return null.  This would frequently cause the
audio to not play since the audio decoder couldn't find a keyframe to
start decoding at.  This didn't affect video decoding since most frames
are not keyframes.

Issue #97
Issue #105
Issue #107

Change-Id: Ia1459862078440585456089cb00fd5fd9b6f12ef