Add TaskRunner::InvokeOrSchedule.
Many of the top-level methods work differently depending on whether
the current thread is the JS main thread.  If called from the JS main
thread, it will invoke the callback directly since it would deadlock
otherwise (waiting for the callback to be invoked).

This adds a helper to TaskRunner to do it there instead.

Change-Id: Ie2d218419f36fe20047e539f13bb0e88fc93fb05