Use Ensure instead of Create/Get in Document.
This allows any thread to create the global document, even before
starting properly.  This will work fine so long as the JsManager exists,
but it may not be started yet.  This is because the BackingObjects
can be created on any thread so long as the ObjectTracker exists to
track their instance.

Change-Id: I28e04ee6419535aa429915458ee884d3b419c67a