Fix retain cycle in ShakaPlayerView.
Using the "target" fields in NSTimer causes the timer to hold a strong
reference to the view.  This means the view can never be destroyed
since the timer is never destroyed.  Now, since we are using a block
with a weak reference, the view will be destroyed once there are no
other references.

Fixes #125

Change-Id: I7105e6215780d336b45599de805de6664d2e64f0