Move common audio renderer code to base class.
This allows more generic testing of the audio renderer and allows us
to write an iOS-specific renderer by sharing the base code.

Before, we used a pull mode where the audio device would request new
data when it needs it.  But this could cause deadlocks since the locks
in our code and the audio device would get acquired in different
orders.  Now we push data into the audio device periodically on a
background thread.  This ensures the locks get acquired in the same
order.

This also removes FFmpeg's swresample since we no longer need it for
resampling or A/V sync.  Now we no longer depend on FFmpeg in the
audio renderer and apps can remove FFmpeg with just removing the
demuxer and the decoder.

Change-Id: Iea93ac3bd7a90054d1f93a2114bce17ee797046d