Fix possible crash in ThreadEvent.
When one thread signals another, the other thread may try to destroy
the ThreadEvent object immediately.  This can cause a crash if the
first thread doesn't unlock the mutex before the object is destroyed.
The second thread gets a copy of the std::future, so there is no
synchronization that prevents it from continuing after the signal is
sent.

So this creates a local copy of the std::promise so the first thread
can send the signal without the lock held.  This allows the object to
be destroyed immediately after signalling.

Change-Id: Ie17a4805f7a24bba71edf87f894e731b1a4aa66e