Fix deadlock in AppleAudioRenderer.
Moving the mutex from the BufferList to the Impl class caused a
deadlock since we wait for the device to get cleared with the lock held,
but the audio device will call our callback on another thread and wait
for the lock.  This moves the lock back and makes the queue_size_
variable atomic since it needs to be thread-safe.

Change-Id: I702d9ed538049d7c62e50c9c54794150d533bc58