Fix render timer when scrolling.
An NSTimer doesn't normally trigger while scrolling, we need to manually
add the timer to the run loop with a different mode to make sure it
fires.

This also changes to using a CADisplayLink for the frame drawing.  This
will be called during the draw phase to ensure we are synchronized with
the device.  Doing this ensures we don't try to update frames between
device refreshes since they won't be seen.

Fixes #135

Change-Id: Ib39786dea89c1a543193ab531e8815c07fb1f78d