Fix lambda capture in UI info.
Before, the lambda had a strong capture of "self", which kept the player
alive.  So if the app removed its reference while waiting for the UI
info callback, the player would get destroyed on the JS main thread and
cause a deadlock.  Now this uses a weak reference so the player gets
destroyed on the app thread.

Fixes #134

Change-Id: Ieecc6112fd991b68925aee2ad5b98346139e6f90