Fix A/V sync for some content.
Instead of syncing each frame to its time, sync to a specific point and
track the bytes written.  This will help long-term A/V sync by avoiding
rounding errors.

This also avoids small sync errors for content when the difference is
small.  This allows playback of some content when there are rounding
errors in the frame times.

Fixes #104

Change-Id: Ib100e32b2c0c69bb53a25960a6ad1438d4d262c5