Remove VideoFillMode::Original.
This mode isn't usable through ShakaPlayerView since it only uses the
Apple enum values, and that isn't an option.  That mode is also not
usable with src= on iOS for the same reason.  Also, this is not really
useful since apps will either have a small or large image; most apps
would want the image to be visually the same for each resolution.

Abi-Breaking: Removes VideoFillMode::Original.
Change-Id: I3764f7b1f05c9fca6293d94a233fc70327d56f3b