Fix possible race in AvMediaPlayer.
First, we should attach the listeners immediately so we can remove them
if the caller calls Detach immediately.  Secondly, if the caller calls
Detach quickly, we need to make sure the background task doesn't run.
Lastly, that background task should acquire the lock too.

We cannot just wait for the background task since it is called on the
JS main thread; so blocking would block that thread.  If the main thread
was also waiting for something to happen on the JS main thread, it would
deadlock.

Change-Id: I986fa82956206808f999abd154bec0003de01148