fix(media): Fix crash on some audio playbacks
Don't assume we have a fill sample.  The platform may remove partial
samples and have a partial sample buffered.  This is mainly used to
determine if we've buffered enough, so a slight inaccuracy is fine.

This also changes the logging for FFmpeg errors to use a macro so the
logs show the correct line numbers.

Lastly, this ignores the nal_ref_idc part of the NALU header when
reading SPS units.

Fixes #174

Cherry-Pick: v1.0.x,v1.x
Change-Id: I00de7f0a5bee780b634335df06738518f9723775